uplink/metainfo: cleanup method names
