+++ uplink/storage/segments/store.go
@@ -138,7 +138,7 @@ func (s *segmentStore) Put(ctx context.Context, streamID storj.StreamID, data io
 		return Meta{}, Error.New("uploaded results (%d) are below the optimal threshold (%d)", l, s.rs.OptimalThreshold())
 	}
 
-	err = s.metainfo.CommitSegmentNew(ctx, metainfo.CommitSegmentParams{
+	err = s.metainfo.CommitSegment(ctx, metainfo.CommitSegmentParams{
 		SegmentID:         segmentID,
 		SizeEncryptedData: sizedReader.Size(),
 		Encryption:        encryption,
